pyproject.toml
[project]
name = "plammindaiintern"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-settings",
    "python-multipart",
    "qdrant-client",
    "redis",
    "langchain-core",
    "langchain-community",
    "langchain-huggingface",
    "sentence-transformers",
    "langchain-text-splitters>=1.1.1",
    "tiktoken>=0.12.0",
    "langchain-openai>=1.1.13",
    "langchain-google-genai>=4.2.2",
    "sqlalchemy>=2.0.49",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "langchain-groq>=0.3.0",
    "pymupdf>=1.24.0",
]
